// web/src/graphql-types.d.ts
export interface PageInfo {
  hasNextPage: boolean;
}
// ...
export interface MarkdownRemark extends Node {
  id: string;
  parent: Node | null;
  children: Array<Node> | null;
  internal: internal_15 | null;
  frontmatter: frontmatter_2 | null;
  rawMarkdownBody: string | null;
  fileAbsolutePath: string | null;
  fields: fields_2 | null;
  html: string | null;
  htmlAst: JSON | null;
  excerpt: string | null;
  headings: Array<MarkdownHeading> | null;
  timeToRead: number | null;
  tableOfContents: string | null;
  wordCount: wordCount | null;
}
// ...
export interface markdownRemarkGroupConnectionConnection {
  pageInfo: PageInfo;
  edges: Array<markdownRemarkGroupConnectionEdge> | null;
  field: string | null;
  fieldValue: string | null;
  totalCount: number | null;
}

export interface markdownRemarkGroupConnectionEdge {
  node: MarkdownRemark | null;
  next: MarkdownRemark | null;
  previous: MarkdownRemark | null;
}